// applets/sonarr/api/model/episodeHistoryEventType.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Sonarr
 * Sonarr API docs
 * OpenAPI spec version: 3.0.0
 */

export type EpisodeHistoryEventType =
  (typeof EpisodeHistoryEventType)[keyof typeof EpisodeHistoryEventType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EpisodeHistoryEventType = {
  unknown: "unknown",
  grabbed: "grabbed",
  seriesFolderImported: "seriesFolderImported",
  downloadFolderImported: "downloadFolderImported",
  downloadFailed: "downloadFailed",
  episodeFileDeleted: "episodeFileDeleted",
  episodeFileRenamed: "episodeFileRenamed",
  downloadIgnored: "downloadIgnored",
} as const;