// applets/sonarr/api/model/historyResource.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Sonarr
 * Sonarr API docs
 * OpenAPI spec version: 3.0.0
 */
import type { CustomFormatResource } from "./customFormatResource";
import type { HistoryResourceData } from "./historyResourceData";
import type { EpisodeResource } from "./episodeResource";
import type { EpisodeHistoryEventType } from "./episodeHistoryEventType";
import type { Language } from "./language";
import type { QualityModel } from "./qualityModel";
import type { SeriesResource } from "./seriesResource";

export interface HistoryResource {
  /** @nullable */
  customFormats?: CustomFormatResource[] | null;
  customFormatScore?: number;
  /** @nullable */
  data?: HistoryResourceData;
  date?: string;
  /** @nullable */
  downloadId?: string | null;
  episode?: EpisodeResource;
  episodeId?: number;
  eventType?: EpisodeHistoryEventType;
  id?: number;
  /** @nullable */
  languages?: Language[] | null;
  quality?: QualityModel;
  qualityCutoffNotMet?: boolean;
  series?: SeriesResource;
  seriesId?: number;
  /** @nullable */
  sourceTitle?: string | null;
}